Add compiler to settings
